Mobile My Jobs account menu + shared Change Password modal
The mobile My Jobs header avatar was a dead circle and techs had no mobile
path to Time Off, notifications, password change, or sign-out. Make the
avatar a real account menu:
- shows the tech's profile photo (falls back to initials)
- menu items: Time Off, Notification settings, Change Password, Sign Out
- reorder the header pills to Pay (first) then Training, per request

Extract ChangePasswordModal into a shared component so the desktop header and
the new mobile menu use one copy instead of duplicating it. Rename the leave
page heading "My Leave" → "My Time Off" and add the /leave route title.
